feat(origin-ui-core): separate OrganizationView from OrganizationForm
